+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Canonical ABI lifting code for `flags` and `variant` types.
+- Lifting code for `result` and `variant` will now panic if caller passes an invalid case.
+- Additional test coverage for `variant` and `flags` cases.
+
+### Removed
+- Package `cm`: Removed unused functions `Reinterpret2`, `LowerResult`, `LowerBool`, `BoolToU64`, `S64ToF64`.
+- Package `cm`: Removed unused experimental `flags` implementation behind a build tag.
+
+### Fixed
+- Lifting code for `result` with no error type will now correctly set `IsErr`.
+
 ## [v0.1.1] — 2024-07-04
 
 This release changes the memory layout of `variant` and `result` types to permit passing these types on the stack safely. This required breaking changes to package `cm`, detailed below, as well as slightly more verbose type signatures for WIT functions that return a typed `result`.